For lisasim-main.cpp I just corrected the include line #include <iostream>
tdifast was broken by changes that Michele made to lisasim-tdi.h.
I have made changes to tdifast so that it is no longer a derived class of
tdi.  It compiles now but the results differ from the output of the standard
version.  There might still be a problem with the units of time.  The old
versions used years, we have now tried to move everything to seconds.

Another possiblity is that Michele updated the tdi version and corrected
something.  I'll look into this later.


git-svn-id: svn+ssh://newman.caltech.edu/newman/user1/vallis/svnarchive/synthlisa/trunk@34 35f09e0b-191c-0410-822b-d9f812da8b64